# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Drachenboot Manager API",
        "version": "2.5.0",
        "description": "API documentation for the Drachenboot Manager application with multi-team support."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/stripe/subscription-details": {
            "get": {
                "summary": "Get Subscription Details",
                "description": "Retrieves detailed subscription information for a team including payment method, billing period, and subscription history.",
                "tags": [
                    "Billing"
                ],
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "query",
                        "required": true,
                        "description": "Team ID to fetch subscription for",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "hasSubscription": {
                                            "type": "boolean"
                                        },
                                        "plan": {
                                            "type": "string",
                                            "description": "Current plan (FREE, PRO)"
                                        },
                                        "isBillingUser": {
                                            "type": "boolean"
                                        },
                                        "isCustomer": {
                                            "type": "boolean"
                                        },
                                        "subscription": {
                                            "type": "object",
                                            "nullable": true,
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "cancelAtPeriodEnd": {
                                                    "type": "boolean"
                                                },
                                                "currentPeriodEnd": {
                                                    "type": "integer"
                                                },
                                                "currentPeriodStart": {
                                                    "type": "integer"
                                                },
                                                "interval": {
                                                    "type": "string",
                                                    "enum": [
                                                        "month",
                                                        "year"
                                                    ]
                                                },
                                                "amount": {
                                                    "type": "integer"
                                                },
                                                "currency": {
                                                    "type": "string"
                                                },
                                                "priceId": {
                                                    "type": "string"
                                                },
                                                "paymentMethod": {
                                                    "type": "object",
                                                    "nullable": true,
                                                    "properties": {
                                                        "brand": {
                                                            "type": "string"
                                                        },
                                                        "last4": {
                                                            "type": "string"
                                                        },
                                                        "expMonth": {
                                                            "type": "integer"
                                                        },
                                                        "expYear": {
                                                            "type": "integer"
                                                        }
                                                    }
                                                }
                                            }
                                        },
                                        "history": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "currentPeriodStart": {
                                                        "type": "integer"
                                                    },
                                                    "currentPeriodEnd": {
                                                        "type": "integer"
                                                    },
                                                    "cancelAtPeriodEnd": {
                                                        "type": "boolean"
                                                    },
                                                    "interval": {
                                                        "type": "string"
                                                    },
                                                    "amount": {
                                                        "type": "integer"
                                                    },
                                                    "currency": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Team ID is required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized or team not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message"
                    }
                },
                "required": [
                    "error"
                ]
            }
        },
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "authjs.session-token",
                "description": "Session-based authentication using NextAuth.js. Requires active login session."
            }
        }
    },
    "security": [
        {
            "sessionAuth": []
        }
    ],
...
}